refactor: Wave-1 dispatch surface — delete dead bytes command (#18), move engine layer out of commands/audio.rs (#21)
Deleted the zero-caller transcribe_audio bytes command (a 4th divergent engine
dispatch) + its handler registration + two orphaned helpers (CloudProvider::transcribe,
save_ai_polish_fallback_history). Moved ActiveEngineSelection, resolve_engine_for_model,
transcribe_whisper_with_acceleration + parakeet helpers + watchdog budget into a new
transcription/engines.rs (with tests); executor.rs now imports from there — zero
crate::commands imports (layering inversion fixed). commands/audio.rs -701 lines,
re-exports keep every callsite unchanged. Pure refactor: 1216 tests pass unchanged, clippy clean.
